Change return type to float, epsilon comparisons

+++ ai-service/src/main/java/tiktok/hackathon/rules/BusinessRule.java
@@ -1,5 +1,5 @@
 package tiktok.hackathon.rules;
 
 public interface BusinessRule {
-  double apply(double prevPrediction);
+  float apply(float prevPrediction);
 }
